Refactor song info page
