+++ src/views/song/Info/comp/SongInfoRelease.tsx
@@ -1,0 +1,20 @@
+/* @refresh skip */
+import { Show } from "solid-js"
+
+import { ReleaseCoverWall } from "~/components/domain/release/ReleaseCoverWall"
+import { assertContext } from "~/utils/context"
+
+import { SongInfoPageContext } from ".."
+
+export function SongInfoRelease() {
+	const ctx = assertContext(SongInfoPageContext)
+	return (
+		<Show when={ctx.song.releases}>
+			{(releases) => (
+				<div class="grid grid-cols-2 gap-4 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6">
+					<ReleaseCoverWall releases={releases()} />
+				</div>
+			)}
+		</Show>
+	)
+}
